requirements.txt
fastapi
uvicorn
python-multipart
llama-index
llama-index-vector-stores-weaviate
llama-index-embeddings-huggingface
llama-index-readers-file
unstructured[pdf]
weaviate-client<4.10.0
loguru
langgraph
langchain
langchain-groq
langchain-core
pydantic
python-dotenv
json-repair
PyJWT
httpx
supabase
cryptography